pull out ads into its own file
